refactor: use index for rexport
